# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  msresponse.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/msresponse.ErrorBody'
      message:
        example: Sucesso
        type: string
      ok:
        example: true
        type: boolean
      request_id:
        example: 7f8c2a9b-5f8a-4b0e-91e5-3b9c2d9a1234
        type: string
    type: object
  msresponse.ErrorBody:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/msresponse.ErrorCode'
        example: 1
      description:
        example: Campo nm_obra é obrigatório
        type: string
    type: object
  msresponse.ErrorCode:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 500
    type: integer
    x-enum-varnames:
    - ErrorFormatoInvalido
    - ErrorTokenInvalido
    - ErrorNaoAutorizado
    - ErrorNaoEncontrado
    - ErrorValidacao
    - ErrorInterno
  user.UserCreate:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
      userrole:
        type: string
    type: object
  user.UserSearch:
    properties:
      nm_search:
        type: string
      nr_docs:
        type: integer
    type: object
  user.UserUpdate:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
      userrole:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://swagger.io
  description: This is a template server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io
  title: Server API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Retorna todos os usuários cadastrados.
      produces:
      - application/json
      responses:
        "200":
          description: Usuários listados com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao listar usuários
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Lista usuários
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Insere um novo usuário na base PostgreSQL. Os campos userrole,
        username, password e email são obrigatórios.
      parameters:
      - description: Dados do usuário a ser cadastrado
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Usuário cadastrado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: JSON inválido ou campos obrigatórios ausentes
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao cadastrar o usuário
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Cadastra um novo usuário
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Exclui um usuário da base PostgreSQL a partir do seu ID numérico.
      parameters:
      - description: ID numérico do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuário removido com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: ID não informado ou inválido
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao remover o usuário
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Remove um usuário por ID
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retorna os dados de um usuário a partir do seu ID numérico.
      parameters:
      - description: ID numérico do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuário localizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: ID não informado ou inválido
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao buscar o usuário
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Busca um usuário por ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuário existente. Se password vier vazio,
        a senha atual será mantida.
      parameters:
      - description: ID numérico do usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do usuário a serem atualizados
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Usuário atualizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: ID inválido ou JSON em formato incorreto
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao atualizar o usuário
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Atualiza um usuário por ID
      tags:
      - users
  /users/email/{email}:
    get:
      consumes:
      - application/json
      description: Retorna os dados de um usuário a partir do email.
      parameters:
      - description: Email do usuário
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuário localizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: Email não informado
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao buscar o usuário
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Busca um usuário por email
      tags:
      - users
  /users/search:
    post:
      consumes:
      - application/json
      description: Realiza busca aproximada de usuários por username, email ou userrole.
      parameters:
      - description: Parâmetros da busca
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/user.UserSearch'
      produces:
      - application/json
      responses:
        "200":
          description: Busca realizada com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: JSON inválido
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao buscar usuários
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Pesquisa usuários
      tags:
      - users
  /users/username/{username}:
    get:
      consumes:
      - application/json
      description: Retorna os dados de um usuário a partir do username.
      parameters:
      - description: Username do usuário
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuário localizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: Username não informado
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao buscar o usuário
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Busca um usuário por username
      tags:
      - users
swagger: "2.0"